refactor: moved to its own hook
