+++ packages/shared/src/hooks/useCookieBanner.ts
@@ -1,19 +1,7 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import { useQuery, useQueryClient } from '@tanstack/react-query';
-import type { CookieOptions } from '../lib/cookie';
-import { expireCookie, getCookies, setCookie } from '../lib/cookie';
+import { useCallback, useEffect, useRef, useState } from 'react';
 import { useAuthContext } from '../contexts/AuthContext';
-import { disabledRefetch } from '../lib/func';
-
-export type AcceptCookiesCallback = (
-  additional?: string[],
-  toRemove?: string[],
-) => void;
-
-interface UseConsentCookie {
-  saveCookies: AcceptCookiesCallback;
-  cookieExists: boolean;
-}
+import type { AcceptCookiesCallback } from './useCookieConsent';
+import { useConsentCookie } from './useCookieConsent';
 
 export const cookieAcknowledgedKey = 'cookie_acknowledged';
 
@@ -41,69 +29,6 @@ export const gdprConsentSettings: Record<GdprConsentKey, ConsentSettings> = {
     description:
       'Marketing cookies are used to deliver content and advertisements that are more relevant to you and your interests. These cookies track your online activity across websites and devices to create a profile of your preferences, enabling personalized experience.',
   },
-};
-
-const cookieOptions: Partial<CookieOptions> = {
-  path: '/',
-  sameSite: 'lax',
-  domain: process.env.NEXT_PUBLIC_DOMAIN,
-};
-
-const setBrowserCookie = (key: string): void => {
-  globalThis?.localStorage.removeItem(key);
-
-  const cookies = getCookies([key]);
-
-  if (cookies[key]) {
-    return;
-  }
-
-  const TEN_YEARS = 60 * 60 * 24 * 365 * 10;
-
-  setCookie(key, true, { ...cookieOptions, maxAge: TEN_YEARS });
-};
-
-const expireBrowserCookie = (key: string): void => {
-  expireCookie(key, { ...cookieOptions });
-  globalThis?.localStorage.setItem(key, 'disabled');
-};
-
-const getBrowserCookie = (key: string) =>
-  globalThis?.document?.cookie.split('; ').find((row) => row.startsWith(key));
-
-export const useConsentCookie = (key: string): UseConsentCookie => {
-  const queryKey = useMemo(() => ['cookie', key], [key]);
-  const client = useQueryClient();
-  const { data: exists } = useQuery({
-    queryKey,
-    initialData: () => !!getBrowserCookie(key),
-    ...disabledRefetch,
-  });
-
-  const onCookieAccepted = useCallback(
-    (cacheKey: string) => {
-      setBrowserCookie(cacheKey);
-      client.setQueryData(['cookie', cacheKey], true);
-    },
-    [client],
-  );
-
-  const saveCookies: AcceptCookiesCallback = useCallback(
-    (additional, toRemove) => {
-      onCookieAccepted(key);
-
-      if (additional) {
-        additional.forEach(onCookieAccepted);
-      }
-
-      if (toRemove) {
-        toRemove.forEach(expireBrowserCookie);
-      }
-    },
-    [key, onCookieAccepted],
-  );
-
-  return { saveCookies, cookieExists: exists };
 };
 
 interface UseCookieBanner {
